Moved pipeline into a package
