+++ scrapy_monkeylearn/pipelines.py
